https://cristianpent12.atlassian.net/jira/software/projects/FG/boards/34?selectedIssue=FG-20
In the Dbintializer change the format a bit and rewrote a lot of it. I created 1 role admin into the DBinitializer with a specialized adminkey. In program Cs updated the database creation to have usermanager and rolemanager

In Accountscontroller updated the Login method so the Admin can log in. Also did a migration